Renaming the files with _jax or _torch / move everything out of folder
