+++ test/plnet_torch/bilipnet_test.py
@@ -1,6 +1,6 @@
 # This file is a part of the RobustNeuralNetworks package. License is MIT: https://github.com/acfr/RobustNeuralNetworks/blob/main/LICENSE 
 
-from robustnn.plnet_torch.bilipnet import BiLipNet
+from robustnn.bilipnet_torch import BiLipNet
 import torch
 import numpy as np
